# Fix file handling to speed up inversion for large-region or long-period runs